+++ src/main/java/model/http/header/HttpCookies.java
@@ -1,4 +1,4 @@
-package model.http.request;
+package model.http.header;
 
 import util.CommonUtils;
